voice_assistant/requirements.txt
# Core dependencies for voice assistant
speechrecognition==3.10.0
openai-whisper==20240930
langchain==0.1.20
langchain-community==0.0.38
ctransformers==0.2.27
gtts==2.4.0
pygame==2.5.2

# Audio dependencies
pyaudio==0.2.11
pydub==0.25.1

# Utility dependencies
requests==2.31.0
numpy==1.24.3
torch==2.0.1
torchaudio==2.0.2

# Optional dependencies for better performance
accelerate==0.21.0
transformers==4.33.2
